global: refactor helper classes and methods into a module
